Group Parameter containers in container module (#1044)
