+++ nevergrad/parametrization/core.py
@@ -6,9 +6,6 @@
 import uuid
 import copy
 import warnings
-import operator
-import functools
-from collections import OrderedDict
 import numpy as np
 import nevergrad.common.typing as tp
 from nevergrad.common import errors
@@ -18,7 +15,6 @@ from . import utils
 
 
 P = tp.TypeVar("P", bound="Parameter")
-D = tp.TypeVar("D", bound="Container")
 X = tp.TypeVar("X")
 
 
@@ -26,6 +22,21 @@ class ValueProperty(tp.Generic[X]):
     """Typed property (descriptor) object so that the value attribute of
     Parameter objects fetches _get_value and _set_value methods
     """
+
+    # This uses the descriptor protocol, like a property:
+    # See https://docs.python.org/3/howto/descriptor.html
+    #
+    # Basically parameter.value calls parameter.value.__get__
+    # and then parameter._get_value
+    def __init__(self) -> None:
+        self.__doc__ = """Value of the Parameter, which should be sent to the function
+        to optimize.
+
+        Example
+        -------
+        >>> ng.p.Array(shape=(2,)).value
+        array([0., 0.])
+        """
 
     def __get__(self, obj: "Parameter", objtype: tp.Optional[tp.Type[object]] = None) -> X:
         return obj._get_value()  # type: ignore
@@ -36,15 +47,21 @@ class ValueProperty(tp.Generic[X]):
 
 # pylint: disable=too-many-instance-attributes,too-many-public-methods
 class Parameter:
-    """Abstract class providing the core functionality of a parameter, aka
+    """Class providing the core functionality of a parameter, aka
     value, internal/model parameters, mutation, recombination
     and additional features such as shared random state,
     constraint check, hashes, generation and naming.
+    The value field should sent to the function to optimize.
 
-    By default, all Parameter attributes of this Parameter are considered as
-    sub-parameters.
-    Spawning a child creates a shallow copy.
+    Example
+    -------
+    >>> ng.p.Array(shape=(2,)).value
+    array([0., 0.])
     """
+
+    # By default, all Parameter attributes of this Parameter are considered as
+    # sub-parameters.
+    # Spawning a child creates a shallow copy.
 
     value: ValueProperty[tp.Any] = ValueProperty()
 
@@ -422,6 +439,9 @@ class Parameter:
         return self._descriptors
 
 
+# Basic types and helpers #
+
+
 class Constant(Parameter):
     """Parameter-like object for simplifying management of constant parameters:
     mutation/recombination do nothing, value cannot be changed, standardize data is an empty array,
@@ -499,146 +519,3 @@ class MultiobjectiveReference(Constant):
                 f"be used by the optimizer.\n(received {parameter} of type {type(parameter)})"
             )
         super().__init__(parameter)
-
-
-class Container(Parameter):
-    """Parameter which can hold other parameters.
-    This abstract implementation is based on a dictionary.
-
-    Parameters
-    ----------
-    **parameters: Any
-        the objects or Parameter which will provide values for the dict
-
-    Note
-    ----
-    This is the base structure for all container Parameters, and it is
-    used to hold the internal/model parameters for all Parameter classes.
-    """
-
-    def __init__(self, **parameters: tp.Any) -> None:
-        super().__init__()
-        self._subobjects = utils.Subobjects(self, base=Parameter, attribute="_content")
-        self._content: tp.Dict[tp.Any, Parameter] = {k: as_parameter(p) for k, p in parameters.items()}
-        self._sizes: tp.Optional[tp.Dict[str, int]] = None
-        self._sanity_check(list(self._content.values()))
-        self._ignore_in_repr: tp.Dict[
-            str, str
-        ] = {}  # hacky undocumented way to bypass boring representations
-
-    def _sanity_check(self, parameters: tp.List[Parameter]) -> None:
-        """Check that all parameters are different"""
-        # TODO: this is first order, in practice we would need to test all the different
-        # parameter levels together
-        if parameters:
-            assert all(isinstance(p, Parameter) for p in parameters)
-            ids = {id(p) for p in parameters}
-            if len(ids) != len(parameters):
-                raise ValueError("Don't repeat twice the same parameter")
-
-    def _compute_descriptors(self) -> utils.Descriptors:
-        init = utils.Descriptors()
-        return functools.reduce(operator.and_, [p.descriptors for p in self._content.values()], init)
-
-    def __getitem__(self, name: tp.Any) -> Parameter:
-        return self._content[name]
-
-    def __len__(self) -> int:
-        return len(self._content)
-
-    def _get_parameters_str(self) -> str:
-        raise NotImplementedError
-
-    def _get_name(self) -> str:
-        return f"{self.__class__.__name__}({self._get_parameters_str()})"
-
-    def get_value_hash(self) -> tp.Hashable:
-        return tuple(sorted((x, y.get_value_hash()) for x, y in self._content.items()))
-
-    def _internal_get_standardized_data(self: D, reference: D) -> np.ndarray:
-        data = {k: self[k].get_standardized_data(reference=p) for k, p in reference._content.items()}
-        if self._sizes is None:
-            self._sizes = OrderedDict(sorted((x, y.size) for x, y in data.items()))
-        assert self._sizes is not None
-        data_list = [data[k] for k in self._sizes]
-        if not data_list:
-            return np.array([])
-        return data_list[0] if len(data_list) == 1 else np.concatenate(data_list)  # type: ignore
-
-    def _internal_set_standardized_data(
-        self: D, data: np.ndarray, reference: D, deterministic: bool = False
-    ) -> None:
-        if self._sizes is None:
-            self.get_standardized_data(reference=self)
-        assert self._sizes is not None
-        if data.size != sum(v for v in self._sizes.values()):
-            raise ValueError(
-                f"Unexpected shape {data.shape} for {self} with dimension {self.dimension}:\n{data}"
-            )
-        data = data.ravel()
-        start, end = 0, 0
-        for name, size in self._sizes.items():
-            end = start + size
-            self._content[name].set_standardized_data(
-                data[start:end], reference=reference[name], deterministic=deterministic
-            )
-            start = end
-        assert end == len(data), f"Finished at {end} but expected {len(data)}"
-
-    def sample(self: D) -> D:
-        child = self.spawn_child()
-        child._content = {k: p.sample() for k, p in self._content.items()}
-        child.heritage["lineage"] = child.uid
-        return child
-
-
-class Dict(Container):
-    """Dictionary-valued parameter. This Parameter can contain other Parameters,
-    its value is a dict, with keys the ones provided as input, and corresponding values are
-    either directly the provided values if they are not Parameter instances, or the value of those
-    Parameters. It also implements a getter to access the Parameters directly if need be.
-
-    Parameters
-    ----------
-    **parameters: Any
-        the objects or Parameter which will provide values for the dict
-
-    Note
-    ----
-    This is the base structure for all container Parameters, and it is
-    used to hold the internal/model parameters for all Parameter classes.
-    """
-
-    def __iter__(self) -> tp.Iterator[str]:
-        return iter(self.keys())
-
-    def keys(self) -> tp.KeysView[str]:
-        return self._content.keys()
-
-    def items(self) -> tp.ItemsView[str, Parameter]:
-        return self._content.items()
-
-    def values(self) -> tp.ValuesView[Parameter]:
-        return self._content.values()
-
-    def _get_value(self) -> tp.Dict[str, tp.Any]:
-        return {k: p.value for k, p in self.items()}
-
-    value: ValueProperty[tp.Dict[str, tp.Any]] = ValueProperty()
-
-    def _set_value(self, value: tp.Dict[str, tp.Any]) -> None:
-        cls = self.__class__.__name__
-        if not isinstance(value, dict):
-            raise TypeError(f"{cls} value must be a dict, got: {value}\nCurrent value: {self.value}")
-        if set(value) != set(self):
-            raise ValueError(
-                f"Got input keys {set(value)} for {cls} but expected {set(self._content)}\nCurrent value: {self.value}"
-            )
-        for key, val in value.items():
-            self._content[key].value = val
-
-    def _get_parameters_str(self) -> str:
-        params = sorted(
-            (k, p.name) for k, p in self.items() if p.name != self._ignore_in_repr.get(k, "#ignoredrepr#")
-        )
-        return ",".join(f"{k}={n}" for k, n in params)
