+++ nevergrad/parametrization/parameter.py
@@ -5,20 +5,28 @@
 
 # pylint: disable=unused-import
 # import with "as" to explicitely allow reexport (mypy)
+
+# abstract types
 from .core import Parameter as Parameter
-from .core import Container as Container  # abstract
-from .core import Dict as Dict
+from .container import Container as Container
+from .choice import BaseChoice as BaseChoice
+from .data import Data as Data
+
+# special types
 from .core import Constant as Constant  # avoid using except for checks
-from .core import (
-    MultiobjectiveReference as MultiobjectiveReference,
-)  # special case for multiobjective optimization
+from .core import MultiobjectiveReference as MultiobjectiveReference  # multiobjective optimization
+
+# containers
+from .container import Dict as Dict
 from .container import Tuple as Tuple
 from .container import Instrumentation as Instrumentation
-from .data import Data as Data  # abstract
+
+# data
 from .data import Array as Array
 from .data import Scalar as Scalar
 from .data import Log as Log
-from .choice import BaseChoice as BaseChoice
+from . import mutation
+
+# choices
 from .choice import Choice as Choice
 from .choice import TransitionChoice as TransitionChoice
-from . import mutation
